+++ src/main/java/local/example/demo/controller/admin/SupplierMgrController.java
@@ -12,8 +12,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import jakarta.validation.Valid;
+import local.example.demo.exception.SupplierInUseException;
 import local.example.demo.model.entity.Product;
 import local.example.demo.model.entity.Supplier;
 import local.example.demo.service.FileService;
@@ -76,6 +78,15 @@ public class SupplierMgrController {
     @PostMapping("save")
     public String saveSupplier(@ModelAttribute("supplier") @Valid Supplier supplier, BindingResult bindingResult,
             @RequestParam("logoFile") MultipartFile logoFile) {
+                
+        // Check for existing supplier name considering create vs update
+        Supplier existingSupplierByName = supplierService.findSupplierByName(supplier.getSupplierName());
+        if (existingSupplierByName != null && 
+            (supplier.getSupplierId() == null || !existingSupplierByName.getSupplierId().equals(supplier.getSupplierId()))) {
+            // Name exists and it's either a new supplier OR it's an existing supplier with a different ID
+            bindingResult.rejectValue("supplierName", "error.supplier", "Tên nhà cung cấp đã tồn tại");
+        }
+        
         if (bindingResult.hasErrors()) {
             return "admin/supplier-mgr/form-supplier";
         }
@@ -90,8 +101,19 @@ public class SupplierMgrController {
 
     // delete the supplier
     @PostMapping("delete/{supplierId}")
-    public String deleteSupplier(@PathVariable("supplierId") Integer supplierId) {
-        supplierService.deleteSupplierById(supplierId);
+    public String deleteSupplier(@PathVariable("supplierId") Integer supplierId, RedirectAttributes redirectAttributes) {
+        Supplier supplier = supplierService.findSupplierById(supplierId);
+        if (supplier != null && supplier.getStatus()) {
+            redirectAttributes.addFlashAttribute("errorMessage", "Không thế xóa nhà cung cấp đang hoạt động");
+            return "redirect:/admin/supplier-mgr/list";
+        }
+        try{
+            supplierService.deleteSupplierById(supplierId);
+            redirectAttributes.addFlashAttribute("successMessage", "Xóa nhà cung cấp thành công");
+        }
+        catch(SupplierInUseException e){
+            redirectAttributes.addFlashAttribute("errorMessage", e.getMessage());
+        }
         return "redirect:/admin/supplier-mgr/list";
     }
